requirements.txt
# Core libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10

# Vector database
faiss-cpu>=1.7.4
chromadb>=0.4.18  # Optional alternative to FAISS

# Embedding models
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.1
unstructured>=0.10.30
nltk>=3.8.1

# LLM connections
langchain-openai>=0.0.2  # If using OpenAI
ollama>=0.1.5  # If using local models

# Utilities
python-dotenv>=1.0.0  # For managing environment variables

# PyMuPDF
PyMuPDF>=1.22.0

# Text parsing libraries
openparse==0.7.0

# PyTorch and related libraries for openparse
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.11.0

# Visualization
streamlit>=1.22.0
pandas>=1.5.0

google-generativeai==0.6.0

# New additions from the code block
llama-index-llms-gemini>=0.1.2
llama-index-vector-stores-qdrant>=0.1.1
llama-index-embeddings-gemini
qdrant-client
llama-index
tiktoken
Pillow
openai